requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.29
alembic==1.13.2
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.1.3
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
python-dateutil==2.9.0
python-dotenv==1.0.1

aiosqlite==0.19.0
email-validator==2.1.0.post1
aiofiles==23.2.1
